# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Campus Second-hand Trading Platform API
  version: 0.1.0
servers:
  - url: http://localhost:7001
paths:
...
  /api/admin/reviews:
    get:
      operationId: listPendingReviews
      summary: 查看待审核商品列表（仅管理员）
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 页码，从 1 开始
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: 每页数量
      responses:
        "200":
          description: 待审核商品列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 仅管理员可执行此操作
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/reviews/{id}/approve:
    post:
      operationId: approveItem
      summary: 审核通过商品（仅管理员）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商品 ID
      responses:
        "200":
          description: 审核通过
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 仅管理员可执行此操作
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 商品不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/reviews/{id}/reject:
    post:
      operationId: rejectItem
      summary: 审核驳回商品（仅管理员）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 商品 ID
      responses:
        "200":
          description: 已驳回
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 仅管理员可执行此操作
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 商品不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/reviews:
    post:
      operationId: createReview
      summary: 提交评价（需认证，仅已签收买家）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: 评价创建成功
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Review"
        "400":
          description: 参数错误或评分无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 无权评价此订单
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 已评价过
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/account:
    get:
      operationId: getAccount
      summary: 查看账户信息（需认证）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 账户信息（余额、是否已设置支付密码）
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/AccountInfo"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/account/balance:
    put:
      operationId: setBalance
      summary: 设置账户余额（演示用途，需认证）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetBalanceInput"
      responses:
        "200":
          description: 余额更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      balance:
                        type: number
        "400":
          description: 无效的余额值
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/account/password:
    put:
      operationId: setPaymentPassword
      summary: 设置支付密码（6 位数字，需认证）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPaymentPasswordInput"
      responses:
        "200":
          description: 支付密码设置成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: 支付密码格式错误（非 6 位数字）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/account/verify-password:
    post:
      operationId: verifyPaymentPassword
      summary: 验证支付密码（需认证）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentPasswordInput"
      responses:
        "200":
          description: 支付密码验证通过
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: 支付密码错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: hex
  schemas:
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    Item:
      type: object
      required:
        [
          id,
          title,
          price,
          quantity,
          description,
          images,
          coverImage,
          category,
          sellerId,
          sellerName,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
        title:
          type: string
        price:
          type: number
          format: float
        quantity:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        coverImage:
          type: string
        category:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        status:
          type: string
          enum: [pending, active, delisted]
        createdAt:
          type: string
          format: date-time
        quantityUpdatedAt:
          type: string
          format: date-time
          nullable: true
        isItemFavorited:
          type: boolean
          description: 当前用户是否收藏该商品（仅登录用户返回有效值）
        isSellerFavorited:
          type: boolean
          description: 当前用户是否收藏该商家（仅登录用户返回有效值）
    ItemListResponse:
      type: object
      required: [data, total, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: integer
        totalPages:
          type: integer
...
    AccountInfo:
      type: object
      required: [balance, hasPaymentPassword]
      properties:
        balance:
          type: number
          description: 账户余额
        hasPaymentPassword:
          type: boolean
          description: 是否已设置支付密码
    SetBalanceInput:
      type: object
      additionalProperties: false
      required: [balance]
      properties:
        balance:
          type: number
          description: 要设置的余额值
    SetPaymentPasswordInput:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password:
          type: string
          pattern: "^\\d{6}$"
          description: 6 位数字支付密码
    VerifyPaymentPasswordInput:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password:
          type: string
          description: 支付密码
...
    Review:
      type: object
      required:
        [
          id,
          userId,
          orderId,
          itemId,
          sellerId,
          rating,
          comment,
          username,
          createdAt,
        ]
      properties:
        id:
          type: integer
        userId:
          type: integer
        orderId:
          type: integer
        itemId:
          type: integer
        sellerId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 10
        comment:
          type: string
        username:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateReviewInput:
      type: object
      additionalProperties: false
      required: [orderId, itemId, rating]
      properties:
        orderId:
          type: integer
        itemId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 10
        comment:
          type: string
          description: 评论文本（可选）
...